Add bunker/water penalty / Tiny bug fixes

+++ Assets/Scripts/States/StrikingState.cs
@@ -11,7 +11,7 @@ public class StrikingState : State
     {
         Mode shotMode = game.GetShotMode().GetShotMode();
         // Hit the damn thing
-        game.GetBall().Strike(shotMode, game.GetBag().GetClub(), (float) game.GetPowerbar().GetPower(), (float) game.GetPowerbar().GetAccuracy());
+        game.GetBall().Strike(shotMode, game.GetBag().GetClub(), game.GetPowerbar().GetPower(), game.GetPowerbar().GetAccuracy());
         game.GetShotMode().Strike();
 
         HoleData hole = game.GetHoleBag().GetCurrentHoleData();
